0.2版 1.不改变原来的文件名 2.优化删除功能，只删除被转移的文件，和空文件夹

+++ src/com/file/copyDemo.java
@@ -18,10 +18,10 @@ public class copyDemo {
         ListAll listAll = new ListAll();
         List<String> list = listAll.ToAllList(FileList,LastName);
         List<String> sonList = listAll.getSonList(FileList);
-        //
-        int i = 0;
+
         for (String name : list){
-            String ToName =FirstName+ (i++)+LastName;
+            File file = new File(name);
+            String ToName =FirstName+ file.getName();
             try (InputStream input = new FileInputStream(name);
                  OutputStream output = new FileOutputStream(ToName))
             {
@@ -32,14 +32,14 @@ public class copyDemo {
         //删除文件
         for (String name : sonList){
             File file = new File(name);
-            deleteFolder(file);
+            deleteFolder(file,LastName);
         }
 
     }
 
     // 删除文件及目录
     //1.传入文件名
-    public static void deleteFolder(File folder) {
+    public static void deleteFolder(File folder,String LastName) {
         //获取文件名下所有文件和文件夹
         File[] files = folder.listFiles();
         //判断文件夹是否为空，为空直接删除
@@ -48,10 +48,14 @@ public class copyDemo {
             for (File f : files) {
                 //判断是否是目录，是则递归，不是删除
                 if (f.isDirectory()) {
-                    deleteFolder(f);
+                    deleteFolder(f,LastName);
                 } else {
-                    //文件，直接删除
-                    f.delete();
+                    //文件,判断是否是要求的格式格式
+                    //定义规则
+                    String rule =  "(\\w*)\\.("+LastName+")";
+                    if (f.getName().matches(rule)){
+                        f.delete();
+                    }
                 }
             }
         }
